requirements.txt
tensorflow-gpu==1.15.0
pytorch
torchvision
tqdm
Cython
imageio